+++ src/index.ts
@@ -3,4 +3,5 @@ export * from "./scalars";
 
 export { buildSchema, BuildSchemaOptions } from "./utils/buildSchema";
 export { useContainer } from "./utils/container";
-export { Resolver } from "./types/decorators";
+export { ArgumentValidationError } from "./errors/ArgumentValidationError";
+export { formatArgumentValidationError } from "./errors/formatArgumentValidationError";