+++ src/resolvers/create.ts
@@ -4,13 +4,20 @@ import { FieldResolverDefinition } from "../metadata/definition-interfaces";
 import { getParams } from "./helpers";
 import { BaseResolverDefinitions } from "../types/resolvers";
 import { convertToType } from "../types/helpers";
+import { BuildContext } from "../schema/build-context";
 
 export function createResolver(
   resolverDefinition: BaseResolverDefinitions,
 ): GraphQLFieldResolver<any, any, any> {
   const targetInstance = IOCContainer.getInstance(resolverDefinition.target);
-  return (root, args, context, info) => {
-    const params = getParams(resolverDefinition.params!, { root, args, context, info });
+  const globalValidate = BuildContext.validate;
+
+  return async (root, args, context, info) => {
+    const params: any[] = await getParams(
+      resolverDefinition.params!,
+      { root, args, context, info },
+      globalValidate,
+    );
     return resolverDefinition.handler!.apply(targetInstance, params);
   };
 }
@@ -23,12 +30,17 @@ export function createFieldResolver(
   }
 
   const targetType = fieldResolverDefintion.getParentType!();
-  return (root, args, context, info) => {
+  return async (root, args, context, info) => {
     const targetInstance: any = convertToType(targetType, root);
+    const globalValidate = BuildContext.validate;
 
     // method
     if (fieldResolverDefintion.handler) {
-      const params = getParams(fieldResolverDefintion.params!, { root, args, context, info });
+      const params: any[] = await getParams(
+        fieldResolverDefintion.params!,
+        { root, args, context, info },
+        globalValidate,
+      );
       return fieldResolverDefintion.handler.apply(targetInstance, params);
     }
     // getter